READme
READme

Description
A fun puzzle game with a modern spin to the retro classic "Where is Waldo". Apo is now Waldo and the stakes are doubled!
Objective
The good guys, known as the Makaveli Squad are trying to protect Apo from the notorious Copyright Villain and they need your help! Apo is being hidden inside one of the boxes try to locate him as many times as you can before the time runs out. 
Instructions
The object of the game is to find Apo. There are two players who will alternate in their chances of finding Apo. The timer is set for 60 seconds. If you successfully locate Apo then you gain a point. Check the scoreboard to see who has found Apo the most before the timer hits zero that player will be declared the winner! As you progress to the next level the grid gets larger and the time decreases. Good Luck!
Player Controls
Use the mouse to click on a box to check if Apo is inside of it. 
Screenshots of finished game
Technologies Used:
HTML, CSS, jQuery, and Javascript.
Resources
Milanote, Invision, MDN, Cole Steel Udemy WebDev Course, Google Search Engine, Google Fonts, Github, W3schools, Wdi_58 notes, Slack, and Visual Studio Code. 
A wireframe of the initial idea. You could take a screenshot of a wireframe 
A Backlog: Descriptions of any unsolved problems or major hurdles you had to overcome.
Bonus: the process mapped out!